removed util func
